Reworked Correlated Request/Response on Core services
When working on support-logging API definition, I realized we needed to
be able to uniquely identify each individual request or response within
the larger batch tied to the correlation ID on the request. According
to the previous definition, each request would have the same value for
correlation ID which is no good if an async caller wants to ensure all
10 requests in a single batch were handled. Each request needs to be
uniquely identified.

Signed-off-by: Trevor Conn <trevor_conn@dell.com>