Mods relative to previously named /bulk endpoint
These changes were made after a conversation with Michael Estrin (Dell)
and were informed by his progress on the POC to show V1/V2 endpoint
side-by-side implementation in the API.

* Rename /bulk to /batch. The latter is clearer.
* Rename BulkRequest to RequestEnvelope
* Rename BulkResponse to ResponseEnvelope
* Both of the above types should have the following properties:
    action, type, version, content, strategy
* The Version property currently on the CorrelatedRequest/Response types
can be moved to RequestEnvelope and ResponseEnvelope specifically. The
API route itself will indicate the major version in an HTTP transport
so there is no need for this property on specific use-case DTOs.

Signed-off-by: Trevor Conn <trevor_conn@dell.com>